separate CartaporteCatalogWrapper from CatalogWrapper
